renamed AnimalWithFilter to Animal, added TrackerAccuracyEvaluator
