+++ src/main/java/sproj/analysis/TrackerAccuracyEvaluator.java
@@ -1,0 +1,4 @@
+package sproj.analysis;
+
+public class TrackerAccuracyEvaluator {
+}
